Move puzzle widget to puzzle tab
